refactor(registry): the wire contract is `@kroma/registry`, not a tools subpath
`@kroma/module-tools/registry` named the contract after the CLI that happens to
emit it. Two of its three consumers are not tooling: the Cloudflare worker serves
these documents and any client parses them, so a publisher reading the import saw
"module tooling" where the answer is "the protocol".

`@kroma/registry` is its own package now, and the dependency points the right way
round: the contract is the leaf (zod and nothing else, worker-safe), and the CLI
depends on it to emit. Semver moved with it - ranges and channels are wire
vocabulary, not build vocabulary - and `Manifest`/`DescribedModule` came along as
the shape a document is built from, leaving `module-tools/catalog.ts` with only
the packer's own `Artifact`/`Entry`/`Catalog`.

No behaviour change; the served paths and documents are byte-identical.
