+++ .github/workflows/repo-worker.yml
@@ -16,10 +16,8 @@ on:
     paths:
       - 'apps/packages/**'
       - 'apps/modules/**'
-      # The registry worker builds the RFC 110 documents with these.
-      - 'packages/module-tools/src/registry/**'
-      - 'packages/module-tools/src/catalog.ts'
-      - 'packages/module-tools/src/semver.ts'
+      # The registry worker builds the RFC 110 documents with this.
+      - 'packages/registry/**'
       - '.github/workflows/repo-worker.yml'
   workflow_dispatch:
 
@@ -49,9 +47,7 @@ jobs:
               - '.github/workflows/repo-worker.yml'
             modules_worker:
               - 'apps/modules/**'
-              - 'packages/module-tools/src/registry/**'
-              - 'packages/module-tools/src/catalog.ts'
-              - 'packages/module-tools/src/semver.ts'
+              - 'packages/registry/**'
               - '.github/workflows/repo-worker.yml'
 
       # Soft gate: no token -> skip (not fail). secrets can't be read in a job-
