+++ packages/module-tools/src/release.ts
@@ -33,9 +33,9 @@
 
 import { existsSync, mkdirSync, readFileSync, writeFileSync } from 'node:fs';
 import { join } from 'node:path';
+import { compareRaw, parse } from '@kroma/registry';
 import { type Bundle, readBundles, toEntries } from './bundles';
 import type { Catalog, Entry } from './catalog';
-import { compareRaw, parse } from './semver';
 import { byCodeUnit } from './sort';
 
 const DEFAULT_REGISTRY = 'https://modules.kroma.tv/modules.json';
