+++ packages/registry/package.json
@@ -1,27 +1,27 @@
 {
-  "name": "@kroma/module-tools",
+  "name": "@kroma/registry",
   "version": "0.0.0",
   "private": true,
   "type": "module",
-  "description": "The module author's CLI: scaffold a module, expand a single-file `*.module.md` into its crate, validate every manifest, pack a `.kmod`, and decide what CI publishes.",
+  "description": "The KROMA module registry contract (RFC 110): the zod schemas every document is parsed through, the builders that emit them, a typed client any conforming registry answers, and the JSON Schema derived from all three.",
   "author": "Maxime Scharwath <maxscharwath@gmail.com>",
   "license": "GPL-2.0-or-later",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/maxscharwath/kroma.git",
-    "directory": "packages/module-tools"
+    "directory": "packages/registry"
   },
+  "main": "src/index.ts",
   "exports": {
-    ".": "./src/cli.ts"
+    ".": "./src/index.ts"
   },
   "scripts": {
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
-    "@kroma/registry": "workspace:*"
+    "zod": "^4.4.3"
   },
   "devDependencies": {
-    "bun-types": "^1.3.14",
     "typescript": "^7.0.2"
   }
 }
