+++ packages/registry/src/client.test.ts
