+++ packages/registry/src/schema.ts
