+++ packages/registry/src/spec.ts
